+++ frontend/src/services/routes.ts
@@ -2,4 +2,7 @@ export const routes = {
   // Buildings
   getBuildings: () => '/buildings',
   getBuilding:  (name: string) => `/buildings/${name}`,
+
+  // Billing
+  getBilling: (buildingName: string) => `/billing/${buildingName}`,
 };